+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - 实现 notify 模块插件以及配套服务。
   - com.dwarfeng.familyhelper.plugin.notify.handler.router.FamilyhelperRouterRegistry。
   - com.dwarfeng.familyhelper.plugin.notify.handler.sender.FamilyhelperSenderRegistry。
-  - com.dwarfeng.familyhelper.plugin.notify.service.SenderExecutorSupportMaintainService。
+  - com.dwarfeng.familyhelper.plugin.notify.service.SendExecutorSupportMaintainService。
   - com.dwarfeng.familyhelper.plugin.notify.service.SyncService。
 
 - 项目建立，清除测试通过。